Task 2.5: Cpp Lint
