+++ include/yolo.hpp
@@ -1,0 +1,364 @@
+/**
+ * @file main.cpp
+ * @author Shantanu Aman Po-Yu 
+ * @brief 
+ * @version 0.1
+ * @date 2022-10-20
+ * 
+ * @copyright Copyright (c) 2022
+ * 
+ */
+#include <fstream>
+#include <sstream>
+#include <iostream>
+
+#include <opencv2/dnn.hpp>
+#include <opencv2/imgproc.hpp>
+#include <opencv2/highgui.hpp>
+#include <chrono>
+
+const char* keys =
+"{help h usage ? | | Usage examples: \n\t\t./object_detection_yolo.out --image=dog.jpg \n\t\t./object_detection_yolo.out --video=run_sm.mp4}"
+"{image i        |<none>| input image   }"
+"{video v       |<none>| input video   }"
+;
+using namespace cv;
+using namespace dnn;
+using namespace std;
+using namespace std::chrono;
+
+// Initialize the parameters
+float confThreshold = 0.5;  // Confidence threshold
+float nmsThreshold = 0.5;  // Non-maximum suppression threshold
+int inpWidth = 416;  // Width of network's input image
+int inpHeight = 416;  // Height of network's input image
+
+vector<string> classes;
+
+// Remove the bounding boxes with low confidence using non-maxima suppression
+void postprocess(Mat& frame, Mat& oldframe, const vector<Mat>& out);
+
+// Draw the predicted bounding box
+void drawPred(int classId, float conf, int left, int top, int right, int bottom, Mat& frame);
+
+// Get the names of the output layers
+vector<String> getOutputsNames(const Net& net);
+
+Size letterbox(Mat frame, int new_shape=416) {
+    // Mat frame = inputframe.clone();
+    int width = frame.cols;
+    int height = frame.rows;
+    // cout<<width << " hain? " << height<< endl;
+    float ratio = float(new_shape)/max(width, height);
+    float ratiow = ratio;
+    float ratioh = ratio;
+    // cout<<width*ratio << " " << height*ratio << endl;
+    int new_unpad0 = int(round(width*ratio));
+    int new_unpad1 = int(round(height * ratio));
+    int dw = ((new_shape - new_unpad0) % 32 )/2;
+    int dh = ((new_shape - new_unpad1) % 32 )/2;
+    int top = int(round(dh - 0.1));
+    int bottom = int(round(dh+0.1));
+    int left = int(round(dw - 0.1));
+    int right = int(round(dw + 0.1));
+
+    // cout<<" ---- "<< new_unpad0 <<  " " << new_unpad1<<endl;
+    cv::resize(frame, frame, cv::Size(new_unpad0, new_unpad1), 0, 0, 1); //CV_INTER_LINEAR = 1
+    Scalar value(127.5, 127.5, 127.5);
+    cv::copyMakeBorder(frame, frame, top, bottom, left, right, cv::BORDER_CONSTANT, value);
+    return frame.size();
+    
+}
+void scale_coords(Size img1, Size img0, Rect &box) {
+    int img00 = img0.height;
+    int img01 = img0.width;
+    int img10 = img1.height;
+    int img11 = img1.width;
+    //  cout<<"im1 ki shape " << img10 << " " << img11 << endl;
+    //  cout<<"im0 ki shape " << img00 << " " << img01 << endl;
+
+    int max0  = max(img00, img01);
+    int max1 = max(img10, img11);
+    double gain = double(max1)/double(max0);
+    // cout<<"Gain = " << gain << " " << max0 << " " << max1 << endl;
+
+    box.x = box.x - (img11 - (img01*gain))/2;
+    box.width = box.width - (img11 - (img01*gain))/2;
+    box.y = box.y - (img10 - (img00*gain))/2;
+    box.height = box.height - (img10 - (img00*gain))/2;
+
+    // cout<<"subtractions = " << box.x << " " << box.y << " " << box.width << " " << box.height << endl;
+
+    box.x = box.x/gain;
+    box.y = box.y/gain;
+    box.width = box.width/gain;
+    box.height = box.height/gain;
+
+    if (box.x < 0)
+        box.x = 0;
+    if (box.y < 0)
+        box.y = 0;
+    if (box.width < 0)
+        box.width = 0;
+    if (box.height < 0)
+        box.height = 0;
+    
+    // cout<<"after gain = " << box.x << " " << box.y << " " << box.width << " " << box.height << endl;
+
+}
+
+int main(int argc, char** argv)
+{
+    CommandLineParser parser(argc, argv, keys);
+    parser.about("Use this script to run object detection using YOLO3 in OpenCV.");
+    if (parser.has("help"))
+    {
+        parser.printMessage();
+        return 0;
+    }
+    // Load names of classes
+    string classesFile = "../cfg/coco.names";
+    ifstream ifs(classesFile.c_str());
+    string line;
+    while (getline(ifs, line)) classes.push_back(line);
+    
+    // Give the configuration and weight files for the model
+    String modelConfiguration = "../cfg/yolov3.cfg";
+    String modelWeights = "../cfg/yolov3.weights";
+
+    // Load the network
+    Net net = readNetFromDarknet(modelConfiguration, modelWeights);
+    net.setPreferableBackend(DNN_BACKEND_OPENCV);
+    net.setPreferableTarget(DNN_TARGET_CPU);
+    
+    // Open a video file or an image file or a camera stream.
+    string str, outputFile;
+    VideoCapture cap;
+    VideoWriter video;
+    Mat frame, blob;
+    
+    try {
+        
+        outputFile = "yolo_out_cpp.avi";
+        if (parser.has("image"))
+        {
+            // Open the image file
+            str = parser.get<String>("image");
+            ifstream ifile(str);
+            if (!ifile) throw("error");
+            cap.open(str);
+            str.replace(str.end()-4, str.end(), "_yolo_out_cpp.jpg");
+            outputFile = str;
+        }
+        else if (parser.has("video"))
+        {
+            // Open the video file
+            str = parser.get<String>("video");
+            ifstream ifile(str);
+            if (!ifile) throw("error");
+            cap.open(str);
+            str.replace(str.end()-4, str.end(), "_yolo_out_cpp.avi");
+            outputFile = str;
+        }
+        // Open the webcaom
+        else cap.open(parser.get<int>("device"));
+        
+    }
+    catch(...) {
+        cout << "Could not open the input image/video stream" << endl;
+        return 0;
+    }
+    
+    // Get the video writer initialized to save the output video
+    if (!parser.has("image")) {
+        video.open(outputFile, VideoWriter::fourcc('M','J','P','G'), 28, Size(cap.get(CAP_PROP_FRAME_WIDTH), cap.get(CAP_PROP_FRAME_HEIGHT)));
+    }
+    
+    // Create a window
+    static const string kWinName = "Deep learning object detection in OpenCV";
+    namedWindow(kWinName, WINDOW_NORMAL);
+    double alltimes = 0.0;
+    double count = 0;
+    Mat oldframe;
+    Size sz;
+
+    // Process frames.
+    while (waitKey(1) < 0)
+    {
+        // get frame from the video
+        cap >> frame;
+
+
+        // Stop the program if reached end of video
+        if (frame.empty()) {
+            cout << "Done processing !!!" << endl;
+            cout << "Output file is stored as " << outputFile << endl;
+            waitKey(3000);
+            break;
+        }
+	    std::chrono::high_resolution_clock::time_point t1 = std::chrono::high_resolution_clock::now();
+        // Create a 4D blob from a frame.
+        cv::cvtColor(frame, frame, COLOR_BGR2RGB);
+
+        // cout<<"Before = "<< frame.size<< " " << frame.cols << " " << frame.rows<< endl;
+        oldframe = frame.clone();
+        sz = letterbox(frame, 416);
+        cout<<"After = "<< sz.width << " " << sz.height << endl;
+        cout<<frame.cols << " " << frame.rows << endl;
+        
+        // blobFromImage(frame, blob, 1/255.0, Size(inpWidth, inpHeight), Scalar(0,0,0), true, false);
+        // blobFromImage(frame, blob, 1/255.0, Size(frame.rows, frame.cols), Scalar(0,0,0), true, false);
+        blobFromImage(frame, blob, 1/255.0, Size(inpWidth, sz.height), Scalar(0,0,0), true, false);
+        
+
+        // cv::FileStorage file("some_name.txt", cv::FileStorage::WRITE);
+
+        // Write to file!
+        // file << "matName" << blob;
+        // cout<<blob.size()<<endl;
+        
+        //Sets the input to the network
+        net.setInput(blob);
+        
+        // Runs the forward pass to get output of the output layers
+        vector<Mat> outs;
+        net.forward(outs, getOutputsNames(net));
+        
+        // Remove the bounding boxes with low confidence
+        postprocess(frame, oldframe, outs);
+	    std::chrono::high_resolution_clock::time_point t2 = std::chrono::high_resolution_clock::now();
+        
+        // Put efficiency information. The function getPerfProfile returns the overall time for inference(t) and the timings for each of the layers(in layersTimes)
+        vector<double> layersTimes;
+        double freq = getTickFrequency() / 1000;
+        double t = net.getPerfProfile(layersTimes) / freq;
+        string label = format("Inference time for a frame : %.2f ms", t);
+        putText(frame, label, Point(0, 15), FONT_HERSHEY_SIMPLEX, 0.5, Scalar(0, 0, 255));
+        double duration = std::chrono::duration_cast<microseconds>( t2 - t1 ).count();
+        alltimes += duration;
+        count +=1;
+
+        std::cout<< label << " chrono time  = " << duration/1000.0<<std::endl;
+            
+        // Write the frame with the detection boxes
+        Mat detectedFrame;
+        frame.convertTo(detectedFrame, CV_8U);
+        cv::cvtColor(detectedFrame,detectedFrame, COLOR_RGB2BGR);
+        if (parser.has("image")) imwrite(outputFile, detectedFrame);
+        else video.write(detectedFrame);
+        
+        imshow(kWinName, frame);
+        
+    }
+    cout<<" MEAN TIME PER FRAME = " << double((alltimes/count)/1000.0) << std::endl;
+    cout<<" TIME PER VIDEO = " << double(alltimes/1000.0) << std::endl;
+    
+    cap.release();
+    if (!parser.has("image")) video.release();
+
+    return 0;
+}
+
+// Remove the bounding boxes with low confidence using non-maxima suppression
+void postprocess(Mat& frame, Mat& oldframe, const vector<Mat>& outs)
+{
+    vector<int> classIds;
+    vector<float> confidences;
+    vector<Rect> boxes;
+    
+    for (size_t i = 0; i < outs.size(); ++i)
+    {
+        // Scan through all the bounding boxes output from the network and keep only the
+        // ones with high confidence scores. Assign the box's class label as the class
+        // with the highest score for the box.
+        float* data = (float*)outs[i].data;
+        for (int j = 0; j < outs[i].rows; ++j, data += outs[i].cols)
+        {
+            Mat scores = outs[i].row(j).colRange(5, outs[i].cols);
+            Point classIdPoint;
+            double confidence;
+            // Get the value and location of the maximum score
+            minMaxLoc(scores, 0, &confidence, 0, &classIdPoint);
+            if (confidence > confThreshold and classIdPoint.x==0)
+            {
+                int centerX = (int)(data[0] * frame.cols);
+                int centerY = (int)(data[1] * frame.rows);
+                int width = (int)(data[2] * frame.cols);
+                int height = (int)(data[3] * frame.rows);
+                int left = centerX - width / 2;
+                int top = centerY - height / 2;
+                
+                classIds.push_back(classIdPoint.x);
+                confidences.push_back((float)confidence); 
+                boxes.push_back(Rect(left, top, width, height));
+                //boxes.push_back(newbox);
+            }
+        }
+    }
+    
+    // Perform non maximum suppression to eliminate redundant overlapping boxes with
+    // lower confidences
+    vector<int> indices;
+    NMSBoxes(boxes, confidences, confThreshold, nmsThreshold, indices);
+    for (size_t i = 0; i < indices.size(); ++i)
+    {
+        int idx = indices[i];
+        Rect box = boxes[idx];
+
+        // box.width = box.x + box.width;
+        // box.height = box.y + box.height;
+
+        // cout<<"Before = " << box.x << " " << box.y << " " << box.width << " " << box.height << endl;
+        
+        // scale_coords(frame.size(), oldframe.size(), box);
+        // cout<<"After = " << box.x << " " << box.y << " " << box.width << " " << box.height << endl;
+
+        drawPred(classIds[idx], confidences[idx], box.x, box.y, box.x + box.width, box.y + box.height, frame);
+        //drawPred(classIds[idx], confidences[idx], box.x, box.y, box.width, box.height, oldframe);
+    }
+}
+
+// Draw the predicted bounding box
+void drawPred(int classId, float conf, int left, int top, int right, int bottom, Mat& frame)
+{
+    //Draw a rectangle displaying the bounding box/
+    
+    // cout<<"draww = " << left << " " << top << " " << right << " " << bottom << endl;
+    rectangle(frame, Point(left, top), Point(right, bottom), Scalar(255, 178, 50), 3);
+    
+    //Get the label for the class name and its confidence
+    string label = format("%.2f", conf);
+    if (!classes.empty() and classId==0)
+    {
+        CV_Assert(classId < (int)classes.size());
+        
+        label = classes[classId] + ":" + label;
+    }
+    
+    //Display the label at the top of the bounding box
+    int baseLine;
+    Size labelSize = getTextSize(label, FONT_HERSHEY_SIMPLEX, 0.5, 1, &baseLine);
+    top = max(top, labelSize.height);
+    rectangle(frame, Point(left, top - round(1.5*labelSize.height)), Point(left + round(1.5*labelSize.width), top + baseLine), Scalar(255, 255, 255), FILLED);
+    putText(frame, label, Point(left, top), FONT_HERSHEY_SIMPLEX, 0.75, Scalar(0,0,0),1);
+}
+
+// Get the names of the output layers
+vector<String> getOutputsNames(const Net& net)
+{
+    static vector<String> names;
+    if (names.empty())
+    {
+        //Get the indices of the output layers, i.e. the layers with unconnected outputs
+        vector<int> outLayers = net.getUnconnectedOutLayers();
+        
+        //get the names of all the layers in the network
+        vector<String> layersNames = net.getLayerNames();
+        
+        // Get the names of the output layers in names
+        names.resize(outLayers.size());
+        for (size_t i = 0; i < outLayers.size(); ++i)
+        names[i] = layersNames[outLayers[i] - 1];
+    }
+    return names;
+}
